Add files via upload
1st TASK: NUMBER GUESSING GAME
Created a C++ program that generates a random number and asks the user to guess it. Provide feedback on whether the guess is too high or too low until the user guesses the correct number. I have used Qt Creater to make the Graphical User Interface (GUI), so that the user can find the game attractive as well as interactive. There are four files I have uploaded which are named as widget.cpp, main.cpp, widget.h and widget.ui.